+++ src/EdgeFieldFactory.jl
@@ -91,7 +91,7 @@ init_field!(_, _) = nothing
 _can_add(_, _, _) = true
 
 
-function construct_edge_functions(T::DataType, attr)
+function construct_edge_functions(T::DataType, attr, simsymbol)
     ignorefrom = :IgnoreFrom in attr[:props]
     singleedge = :SingleEdge in attr[:props]
     singletype = :SingleAgentType in attr[:props]
@@ -170,8 +170,8 @@ function construct_edge_functions(T::DataType, attr)
                               pointer(field, s+1), 0, (nr-s) * sizeof($CT))
                     end
                 end
-            else 
-                @eval function init_field!(sim, ::Val{$MT})
+            else
+                @eval function init_field!(sim::$simsymbol, ::Val{$MT})
                     field = sim.$(writefield(T))
                     resize!(field, $singletype_size)
                     ccall(:memset, Nothing, (Ptr{Int64}, Int8, Int64),
@@ -187,7 +187,7 @@ function construct_edge_functions(T::DataType, attr)
                     end
                 end
             else
-                @eval init_field!(sim, ::Val{$MT}) =
+                @eval init_field!(sim::$simsymbol, ::Val{$MT}) =
                     resize!(sim.$(writefield(T)), $singletype_size)
             end
             
@@ -208,7 +208,7 @@ function construct_edge_functions(T::DataType, attr)
 
     #- _get_agent_container
     if singletype
-        @eval function _get_agent_container(sim, to::AgentID, ::Type{$(Val{T})}, field)
+        @eval function _get_agent_container(sim::$simsymbol, to::AgentID, ::Type{$(Val{T})}, field)
             nr = _to2idx(to, $(Val{T}))
 
             @mayassert begin
@@ -225,7 +225,7 @@ function construct_edge_functions(T::DataType, attr)
             field[nr]
         end
     else
-        @eval function _get_agent_container(sim, to::AgentID, ::Type{$(Val{T})}, field)
+        @eval function _get_agent_container(::$simsymbol, to::AgentID, ::Type{$(Val{T})}, field)
             get!(_construct_container_func($CT), field, to)
         end
     end
@@ -233,19 +233,19 @@ function construct_edge_functions(T::DataType, attr)
     #### The exported functions
     #- add_edge!
     if stateless && ignorefrom && !singleedge
-        @eval function add_edge!(sim::Simulation, to::AgentID, ::Edge{$MT})
+        @eval function add_edge!(sim::$simsymbol, to::AgentID, ::Edge{$MT})
             nr = _to2idx(to, $(Val{T}))
             sim.$(writefield(T))[nr] =
                 _get_agent_container(sim, to, $(Val{T}), sim.$(writefield(T))) + 1
         end
 
-        @eval function add_edge!(sim::Simulation, ::AgentID, to::AgentID, ::$MT)
+        @eval function add_edge!(sim::$simsymbol, ::AgentID, to::AgentID, ::$MT)
             nr = _to2idx(to, $(Val{T}))
             sim.$(writefield(T))[nr] =
                 _get_agent_container(sim, to, $(Val{T}), sim.$(writefield(T))) + 1
         end
     elseif singleedge
-        @eval function add_edge!(sim::Simulation, to::AgentID, edge::Edge{$MT})
+        @eval function add_edge!(sim::$simsymbol, to::AgentID, edge::Edge{$MT})
             @mayassert _can_add(sim.$(writefield(T)), to, $(Val{T})) """
             An edge has already beend added to the agent with the id $to (and the
             edgetype properties containing the :SingleEdge property).
@@ -255,7 +255,7 @@ function construct_edge_functions(T::DataType, attr)
             sim.$(writefield(T))[nr] = _valuetostore(edge)
         end
 
-        @eval function add_edge!(sim::Simulation, from::AgentID, to::AgentID, edgestate::$MT)
+        @eval function add_edge!(sim::$simsymbol, from::AgentID, to::AgentID, edgestate::$MT)
             @mayassert _can_add(sim.$(writefield(T)), to, $(Val{T})) """
             An edge has already beend added to the agent with the id $to (and the
             edgetype properties containing the :SingleEdge property).
@@ -265,25 +265,25 @@ function construct_edge_functions(T::DataType, attr)
             sim.$(writefield(T))[nr] = _valuetostore(from, edgestate)
         end
     else
-        @eval function add_edge!(sim::Simulation, to::AgentID, edge::Edge{$MT})
+        @eval function add_edge!(sim::$simsymbol, to::AgentID, edge::Edge{$MT})
             push!(_get_agent_container(sim, to, $(Val{T}), sim.$(writefield(T))),
                   _valuetostore(edge))
         end
 
-        @eval function add_edge!(sim::Simulation, from::AgentID, to::AgentID, edgestate::$MT)
+        @eval function add_edge!(sim::$simsymbol, from::AgentID, to::AgentID, edgestate::$MT)
             push!(_get_agent_container(sim, to, $(Val{T}), sim.$(writefield(T))),
                   _valuetostore(from, edgestate))
         end
     end
 
     #- prepare_write! 
-    @eval function prepare_write!(sim, t::Val{$MT})
+    @eval function prepare_write!(sim::$simsymbol, t::Val{$MT})
         sim.$(writefield(T)) = $FT()
         init_field!(sim, t)
     end
 
     #- finish_write!
-    @eval function finish_write!(sim, ::Val{$MT})
+    @eval function finish_write!(sim::$simsymbol, ::Val{$MT})
         sim.$(readfield(T)) = sim.$(writefield(T))
     end
 
@@ -294,11 +294,11 @@ function construct_edge_functions(T::DataType, attr)
 
     #- edges_to
     if !stateless && !ignorefrom
-        @eval function edges_to(sim::Simulation, to::AgentID, ::Val{$MT})
+        @eval function edges_to(sim::$simsymbol, to::AgentID, ::Val{$MT})
             _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T)))
         end
     else
-        @eval function edges_to(::Simulation, ::AgentID, t::Val{$MT})
+        @eval function edges_to(::$simsymbol, ::AgentID, t::Val{$MT})
             @assert false """
             edges_to is not defined for the property combination of $t
             """
@@ -308,23 +308,23 @@ function construct_edge_functions(T::DataType, attr)
     #- neighborids
     if !ignorefrom
         if stateless
-            @eval function neighborids(sim::Simulation, to::AgentID, ::Val{$MT})
+            @eval function neighborids(sim::$simsymbol, to::AgentID, ::Val{$MT})
                 _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T)))
             end
         else
             if singleedge
-                @eval function neighborids(sim::Simulation, to::AgentID, ::Val{$MT})
+                @eval function neighborids(sim::$simsymbol, to::AgentID, ::Val{$MT})
                     _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))).from
                 end
             else
-                @eval function neighborids(sim::Simulation, to::AgentID, ::Val{$MT})
+                @eval function neighborids(sim::$simsymbol, to::AgentID, ::Val{$MT})
                     map(e -> e.from,
                         _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))))
                 end
             end
         end
     else
-        @eval function neighborids(::Simulation, ::AgentID, t::Val{$MT})
+        @eval function neighborids(::$simsymbol, ::AgentID, t::Val{$MT})
             @assert false """
             neighborids is not defined for the property combination of $t
             """
@@ -334,23 +334,23 @@ function construct_edge_functions(T::DataType, attr)
     #- edgestates
     if !stateless
         if ignorefrom
-            @eval function edgestates(sim::Simulation, to::AgentID, ::Val{$MT})
+            @eval function edgestates(sim::$simsymbol, to::AgentID, ::Val{$MT})
                 _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T)))
             end
         else
             if singleedge
-                @eval function edgestates(sim::Simulation, to::AgentID, ::Val{$MT})
+                @eval function edgestates(sim::$simsymbol, to::AgentID, ::Val{$MT})
                     _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))).state
                 end
             else
-                @eval function edgestates(sim::Simulation, to::AgentID, ::Val{$MT})
+                @eval function edgestates(sim::$simsymbol, to::AgentID, ::Val{$MT})
                     map(e -> e.state,
                         _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))))
                 end
             end
         end
     else
-        @eval function edgestates(::Simulation, ::AgentID, t::Val{$MT})
+        @eval function edgestates(::$simsymbol, ::AgentID, t::Val{$MT})
             @assert false """
             edgestates is not defined for the property combination of $t
             """
@@ -360,16 +360,16 @@ function construct_edge_functions(T::DataType, attr)
     #- num_neighbors
     if !singleedge
         if ignorefrom && stateless
-            @eval function num_neighbors(sim::Simulation, to::AgentID, ::Val{$MT})
+            @eval function num_neighbors(sim::$simsymbol, to::AgentID, ::Val{$MT})
                 _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T)))
             end
         else
-            @eval function num_neighbors(sim::Simulation, to::AgentID, ::Val{$MT})
+            @eval function num_neighbors(sim::$simsymbol, to::AgentID, ::Val{$MT})
                 size(_get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))),1)
             end
         end
     else
-        @eval function num_neighbors(::Simulation, ::AgentID, t::Val{$MT})
+        @eval function num_neighbors(::$simsymbol, ::AgentID, t::Val{$MT})
             @assert false """
             num_neighbors is not defined for the property combination of $t
             """
@@ -379,19 +379,19 @@ function construct_edge_functions(T::DataType, attr)
 
     #- has_neighbor
     if ignorefrom && stateless
-        @eval function has_neighbor(sim::Simulation, to::AgentID, ::Val{$MT})
+        @eval function has_neighbor(sim::$simsymbol, to::AgentID, ::Val{$MT})
             _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))) >= 1
         end
     elseif !singleedge
-        @eval function has_neighbor(sim::Simulation, to::AgentID, t::Val{$MT})
+        @eval function has_neighbor(sim::$simsymbol, to::AgentID, t::Val{$MT})
             num_neighbors(sim,to, t) >= 1
         end
     elseif !singletype 
-        @eval function has_neighbor(sim::Simulation, to::AgentID, ::Val{$MT})
+        @eval function has_neighbor(sim::$simsymbol, to::AgentID, ::Val{$MT})
             haskey(sim.$(readfield(T)), to)
         end
     else
-        @eval function has_neighbor(::Simulation, ::AgentID, t::Val{$MT})
+        @eval function has_neighbor(::$simsymbol, ::AgentID, t::Val{$MT})
             @assert false """
             has_neighbor is not defined for the property combination of $t
             """
@@ -400,17 +400,17 @@ function construct_edge_functions(T::DataType, attr)
 
     #- _vectorized_container
     # if !singleedge
-    #     @eval function _vectorized_container_read(sim::Simulation, to::AgentID, ::Val{$MT})
+    #     @eval function _vectorized_container_read(sim::$simsymbol, to::AgentID, ::Val{$MT})
     #         _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))) 
     #     end
-    #     @eval function _vectorized_container_write(sim::Simulation, to::AgentID, ::Val{$MT})
+    #     @eval function _vectorized_container_write(sim::$simsymbol, to::AgentID, ::Val{$MT})
     #         _get_agent_container(sim, to, $(Val{T}), sim.$(writefield(T))) 
     #     end
     # else
-    #     @eval function _vectorized_container_read(sim::Simulation, to::AgentID, ::Val{$MT})
+    #     @eval function _vectorized_container_read(sim::$simsymbol, to::AgentID, ::Val{$MT})
     #         [ _get_agent_container(sim, to, $(Val{T}), sim.$(readfield(T))) ]
     #     end
-    #     @eval function _vectorized_container_write(sim::Simulation, to::AgentID, ::Val{$MT})
+    #     @eval function _vectorized_container_write(sim::$simsymbol, to::AgentID, ::Val{$MT})
     #         [ _get_agent_container(sim, to, $(Val{T}), sim.$(writefield(T))) ]
     #     end
     # end
@@ -425,7 +425,7 @@ function construct_edge_functions(T::DataType, attr)
 
     if !singleedge
         if !singletype
-            @eval function _num_edges(sim::Simulation, t::Val{$MT}, write = false)
+            @eval function _num_edges(sim::$simsymbol, t::Val{$MT}, write = false)
                 field = write ? sim.$(writefield(T)) : sim.$(readfield(T))
                 if length(field) == 0
                     return 0
@@ -433,7 +433,7 @@ function construct_edge_functions(T::DataType, attr)
                 mapreduce(id -> length(field[id]), +, keys(field))
             end
         else
-            @eval function _num_edges(sim::Simulation, t::Val{$MT}, write = false)
+            @eval function _num_edges(sim::$simsymbol, t::Val{$MT}, write = false)
                 field = write ? sim.$(writefield(T)) : sim.$(readfield(T))
                 if length(field) == 0
                     return 0
@@ -448,7 +448,7 @@ function construct_edge_functions(T::DataType, attr)
             end
         end 
     else
-        @eval function _num_edges(sim::Simulation, t::Val{$MT}, write = false)
+        @eval function _num_edges(sim::$simsymbol, t::Val{$MT}, write = false)
             field = write ? sim.$(writefield(T)) : sim.$(readfield(T))
             # TODO: performance improvement with _countundef?
             length(_removeundef(t)(field))
@@ -463,7 +463,7 @@ function construct_edge_functions(T::DataType, attr)
             @eval _edgestates(::Val{$MT}) = edges -> map(e -> e.state, edges)
         end
         if singleedge
-            @eval function aggregate(sim::Simulation, t::Val{$MT}, f, op; kwargs...)
+            @eval function aggregate(sim::$simsymbol, t::Val{$MT}, f, op; kwargs...)
                 estates = sim.$(readfield(T)) |>
                     values |>
                     _removeundef(t) |>
@@ -472,7 +472,7 @@ function construct_edge_functions(T::DataType, attr)
                 mapreduce(f, op, estates; kwargs...)
             end
         else
-            @eval function aggregate(sim::Simulation, t::Val{$MT}, f, op; kwargs...)
+            @eval function aggregate(sim::$simsymbol, t::Val{$MT}, f, op; kwargs...)
                 estates = sim.$(readfield(T)) |>
                     values |>
                     _removeundef(t) |>
@@ -483,7 +483,7 @@ function construct_edge_functions(T::DataType, attr)
             end
         end
     else
-        @eval function aggregate(::Simulation, t::Val{$MT}, f, op; kwargs...)
+        @eval function aggregate(::$simsymbol, t::Val{$MT}, f, op; kwargs...)
             @assert false """
             aggregate is not defined for the property combination of $t
             """